Extract StyleElements from TrackSchemeStyleEditor
