+++ src/main/java/org/mastodon/revised/trackscheme/display/style/TrackSchemeStyleEditorPanel.java
@@ -2,28 +2,17 @@ package org.mastodon.revised.trackscheme.display.style;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
-import java.awt.Component;
-import java.awt.Graphics;
-import java.awt.Graphics2D;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
-import java.awt.RenderingHints;
 import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
-import java.awt.geom.RoundRectangle2D;
 import java.util.Arrays;
 import java.util.List;
-import java.util.function.BooleanSupplier;
-import java.util.function.Consumer;
-import java.util.function.Supplier;
-
 import javax.swing.AbstractAction;
 import javax.swing.Action;
 import javax.swing.ActionMap;
 import javax.swing.Box;
-import javax.swing.Icon;
 import javax.swing.InputMap;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
@@ -32,15 +21,17 @@ import javax.swing.JComponent;
 import javax.swing.JDialog;
 import javax.swing.JPanel;
 import javax.swing.KeyStroke;
-import javax.swing.SwingConstants;
 import javax.swing.WindowConstants;
-import javax.swing.border.EmptyBorder;
 import javax.swing.border.LineBorder;
-
 import org.mastodon.adapter.FocusModelAdapter;
 import org.mastodon.adapter.HighlightModelAdapter;
 import org.mastodon.adapter.RefBimap;
 import org.mastodon.adapter.SelectionModelAdapter;
+import org.mastodon.app.ui.settings.StyleElements.BooleanSetter;
+import org.mastodon.app.ui.settings.StyleElements.ColorSetter;
+import org.mastodon.app.ui.settings.StyleElements.Separator;
+import org.mastodon.app.ui.settings.StyleElements.StyleElement;
+import org.mastodon.app.ui.settings.StyleElements.StyleElementVisitor;
 import org.mastodon.graph.GraphIdBimap;
 import org.mastodon.model.DefaultFocusModel;
 import org.mastodon.model.DefaultHighlightModel;
@@ -63,6 +54,13 @@ import org.mastodon.revised.trackscheme.display.style.dummygraph.DummyGraph;
 import org.mastodon.revised.trackscheme.display.style.dummygraph.DummyVertex;
 import org.mastodon.revised.trackscheme.wrap.DefaultModelGraphProperties;
 import org.mastodon.revised.trackscheme.wrap.ModelGraphProperties;
+
+import static org.mastodon.app.ui.settings.StyleElements.booleanSetter;
+import static org.mastodon.app.ui.settings.StyleElements.colorSetter;
+import static org.mastodon.app.ui.settings.StyleElements.linkedCheckBox;
+import static org.mastodon.app.ui.settings.StyleElements.linkedColorButton;
+import static org.mastodon.app.ui.settings.StyleElements.separator;
+
 
 public class TrackSchemeStyleEditorPanel extends JPanel
 {
@@ -129,193 +127,53 @@ public class TrackSchemeStyleEditorPanel extends JPanel
 		c.gridwidth = 1;
 
 		c.gridx = 0;
-		for ( final StyleElement element : styleElements )
-		{
-			if ( element.fillsRow() )
-			{
-				if ( c.gridx != 0 )
+
+		styleElements.forEach( element -> element.accept(
+				new StyleElementVisitor()
 				{
-					c.gridx = 0;
-					++c.gridy;
+					@Override
+					public void visit( final Separator separator )
+					{
+						if ( c.gridx != 0 )
+						{
+							c.gridx = 0;
+							++c.gridy;
+						}
+						c.gridwidth = 2;
+						editPanel.add( Box.createVerticalStrut( 10 ), c );
+						++c.gridy;
+						c.gridwidth = 1;
+					}
+
+					@Override
+					public void visit( final ColorSetter element )
+					{
+						JButton button = linkedColorButton( element, element.getLabel(), colorChooser );
+						editPanel.add( button, c );
+						if ( ++c.gridx == numCols )
+						{
+							c.gridx = 0;
+							++c.gridy;
+						}
+					}
+
+					@Override
+					public void visit( final BooleanSetter element )
+					{
+						JCheckBox checkbox = linkedCheckBox( element, element.getLabel() );
+						editPanel.add( checkbox, c );
+						if ( ++c.gridx == numCols )
+						{
+							c.gridx = 0;
+							++c.gridy;
+						}
+					}
 				}
-				c.gridwidth = 2;
-				editPanel.add( element.getComponent(), c );
-				++c.gridy;
-				c.gridwidth = 1;
-			}
-			else
-			{
-				editPanel.add( element.getComponent(), c );
-				if ( ++c.gridx == numCols )
-				{
-					c.gridx = 0;
-					++c.gridy;
-				}
-			}
-		}
+		) );
 
 		previewPanel.setBorder( new LineBorder( Color.LIGHT_GRAY, 1 ) );
 		add( previewPanel, BorderLayout.CENTER );
 		add( editPanel, BorderLayout.SOUTH );
-	}
-
-	private interface StyleElement
-	{
-		public Component getComponent();
-
-		public default void update()
-		{}
-
-		public default boolean fillsRow()
-		{
-			return false;
-		}
-	}
-
-	private class Separator implements StyleElement
-	{
-		@Override
-		public Component getComponent()
-		{
-			return Box.createVerticalStrut( 10 );
-		}
-
-		@Override
-		public boolean fillsRow()
-		{
-			return true;
-		}
-	}
-
-	private Separator separator()
-	{
-		return new Separator();
-	}
-
-	private abstract class ColorSetter implements StyleElement
-	{
-		private final ColorIcon icon;
-
-		private final JButton button;
-
-		public ColorSetter( final String label )
-		{
-			icon = new ColorIcon( getColor() );
-			button = new JButton( label, icon );
-			button.setOpaque( false );
-			button.setContentAreaFilled( false );
-			button.setBorderPainted( false );
-			button.setFont( new JButton().getFont() );
-			button.setMargin( new Insets( 0, 0, 0, 0 ) );
-			button.setBorder( new EmptyBorder( 2, 5, 2, 2 ) );
-			button.setHorizontalAlignment( SwingConstants.LEFT );
-			button.setFocusable( false );
-			button.addActionListener( new ActionListener()
-			{
-				@Override
-				public void actionPerformed( final ActionEvent e )
-				{
-					colorChooser.setColor( getColor() );
-					final JDialog d = JColorChooser.createDialog( button, "Choose a color", true, colorChooser, new ActionListener()
-					{
-						@Override
-						public void actionPerformed( final ActionEvent arg0 )
-						{
-							final Color c = colorChooser.getColor();
-							if ( c != null )
-							{
-								icon.setColor( c );
-								button.repaint();
-								setColor( c );
-							}
-						}
-					}, null );
-					d.setVisible( true );
-				}
-			} );
-		}
-
-		public abstract Color getColor();
-
-		public abstract void setColor( Color c );
-
-		@Override
-		public JButton getComponent()
-		{
-			return button;
-		}
-
-		@Override
-		public void update()
-		{
-//			if ( ! icon.color.equals( getColor() ) )
-				icon.setColor( getColor() );
-		}
-	}
-
-	private ColorSetter colorSetter( final String label, final Supplier< Color > get, final Consumer< Color > set )
-	{
-		return new ColorSetter( label )
-		{
-			@Override
-			public Color getColor()
-			{
-				return get.get();
-			}
-
-			@Override
-			public void setColor( final Color c )
-			{
-				set.accept( c );
-			}
-		};
-	}
-
-	private abstract class BooleanSetter implements StyleElement
-	{
-		private final JCheckBox checkbox;
-
-		public BooleanSetter( final String label )
-		{
-			checkbox = new JCheckBox( label, get() );
-			checkbox.setFocusable( false );
-			checkbox.addActionListener( ( e ) -> set( checkbox.isSelected() ) );
-		}
-
-		public abstract boolean get();
-
-		public abstract void set( boolean b );
-
-		@Override
-		public JCheckBox getComponent()
-		{
-			return checkbox;
-		}
-
-		@Override
-		public void update()
-		{
-			if ( get() != checkbox.isSelected() )
-				checkbox.setSelected( get() );
-		}
-	}
-
-	private BooleanSetter booleanSetter( final String label, final BooleanSupplier get, final Consumer< Boolean > set )
-	{
-		return new BooleanSetter( label )
-		{
-			@Override
-			public boolean get()
-			{
-				return get.getAsBoolean();
-			}
-
-			@Override
-			public void set( final boolean b )
-			{
-				set.accept( b );
-			}
-		};
 	}
 
 	private List< StyleElement > styleElements( final TrackSchemeStyle style )
@@ -350,50 +208,6 @@ public class TrackSchemeStyleEditorPanel extends JPanel
 				booleanSetter( "paint columns", style::isPaintColumns, style::paintColumns ),
 				booleanSetter( "paint header shadow", style::isPaintHeaderShadow, style::paintHeaderShadow )
 		);
-	}
-
-	/**
-	 * Adapted from http://stackoverflow.com/a/3072979/230513
-	 */
-	private static class ColorIcon implements Icon
-	{
-		private final int size = 16;
-
-		private final int pad = 2;
-
-		private Color color;
-
-		public ColorIcon( final Color color )
-		{
-			this.color = color;
-		}
-
-		@Override
-		public void paintIcon( final Component c, final Graphics g, final int x, final int y )
-		{
-			final Graphics2D g2d = ( Graphics2D ) g;
-			g2d.setRenderingHint( RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON );
-			g2d.setColor( color );
-			// g2d.fillOval( x, y, size, size );
-			g2d.fill( new RoundRectangle2D.Float( x + pad, y + pad, size, size, 5, 5 ) );
-		}
-
-		public void setColor( final Color color )
-		{
-			this.color = color;
-		}
-
-		@Override
-		public int getIconWidth()
-		{
-			return size + 2 * pad;
-		}
-
-		@Override
-		public int getIconHeight()
-		{
-			return size + 2 * pad;
-		}
 	}
 
 	public static void main( final String[] args )
